refactor(yakbak-server): rename to "proxies"
